requirements.txt
# Core Dependencies
Django>=3.2.0,<4.0.0
channels>=3.0.0
channels-redis>=3.3.0
stripe>=2.60.0
xhtml2pdf>=0.2.5
django-filter>=2.4.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=4.8.0
Pillow>=8.2.0
django-cors-headers>=3.7.0
django-storages>=1.11.1
django-redis>=5.2.0
psycopg2-binary>=2.9.1
gunicorn>=20.1.0
python-dotenv>=0.19.0
drf-yasg>=1.20.0
drf-spectacular>=0.28.0

# Task Queue
celery>=5.1.2
redis>=4.2.0

# Testing and Development
pytest>=6.2.5
pytest-django>=4.4.0
pytest-cov>=2.12.1
black>=21.6b0
flake8>=3.9.2
isort>=5.9.2
django-debug-toolbar>=3.2.2
sentry-sdk>=1.3.1

# Data Science and Forecasting
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
statsmodels>=0.13.2
prophet>=1.1.1
scipy>=1.7.0

# Monitoring and Logging
python-json-logger>=2.0.0
django-prometheus>=2.1.0
prometheus-client>=0.12.0

# Documentation
mkdocs>=1.2.3
mkdocs-material>=8.0.0
mkdocstrings>=0.16.0